+++ test/Feature/CleanRegex/PcrePatternTest.php
@@ -31,7 +31,7 @@ class PcrePatternTest extends TestCase
         // when
         $pattern = PcrePattern::template('%You/her, @ (her)%s')->mask('%s', ['%s' => '\s']);
         // then
-        $this->assertPatternIs('%You/her, \s (her)%s', $pattern);
+        $this->assertPatternIs('%You/her, (?:\s) (her)%s', $pattern);
     }
 
     /**
@@ -42,7 +42,7 @@ class PcrePatternTest extends TestCase
         // when
         $pattern = PcrePattern::template('@You/her, @ (her)@s')->literal('{hi@}');
         // then
-        $this->assertPatternIs('@You/her, \{hi\@\} (her)@s', $pattern);
+        $this->assertPatternIs('@You/her, (?>\{hi\@\}) (her)@s', $pattern);
     }
 
     /**