+++ test/Feature/CleanRegex/_extended/_comments/builder/PatternTest.php
@@ -18,7 +18,7 @@ class PatternTest extends TestCase
         $pattern = Pattern::builder("/#@\n", 'i')->literal('cat~')->build();
         // when, then
         $this->assertConsumesFirst("/#cat~\n", $pattern);
-        $this->assertPatternIs("%/#cat~\n%i", $pattern);
+        $this->assertPatternIs("%/#(?>cat~)\n%i", $pattern);
     }
 
     /**