+++ test/Feature/CleanRegex/_figures/_backtracking/PatternTest.php
@@ -12,6 +12,29 @@ class PatternTest extends TestCase
     /**
      * @test
      */
+    public function shouldAllowBacktrackingMask()
+    {
+        // given
+        $pattern = Pattern::mask('*Bar', ['*' => 'FooBar|Foo']);
+        // when, then
+        $this->assertConsumesFirst('FooBar', $pattern);
+    }
+
+    /**
+     * @test
+     */
+    public function shouldAllowBacktrackingTemplateMask()
+    {
+        // given
+        $pattern = Pattern::template('@Bar')->mask('*', ['*' => 'FooBar|Foo']);
+        // when, then
+        $this->assertConsumesFirst('FooBar', $pattern);
+        $this->assertConsumesFirst('FooBarBar', $pattern);
+    }
+
+    /**
+     * @test
+     */
     public function shouldAllowBacktrackingTemplateAlteration()
     {
         // given
@@ -24,10 +47,46 @@ class PatternTest extends TestCase
     /**
      * @test
      */
+    public function shouldAllowBacktrackingTemplatePattern()
+    {
+        // given
+        $pattern = Pattern::template('@Bar')->pattern('FooBar|Foo');
+        // when, then
+        $this->assertConsumesFirst('FooBar', $pattern);
+        $this->assertConsumesFirst('FooBarBar', $pattern);
+    }
+
+    /**
+     * @test
+     */
+    public function shouldAllowBacktrackingBuilderTemplateMask()
+    {
+        // given
+        $pattern = Pattern::builder('@Bar')->mask('*', ['*' => 'FooBar|Foo'])->build();
+        // when, then
+        $this->assertConsumesFirst('FooBar', $pattern);
+        $this->assertConsumesFirst('FooBarBar', $pattern);
+    }
+
+    /**
+     * @test
+     */
     public function shouldAllowBacktrackingBuilderTemplateAlteration()
     {
         // given
         $pattern = Pattern::builder('@Bar')->alteration(['FooBar', 'Foo'])->build();
+        // when, then
+        $this->assertConsumesFirst('FooBar', $pattern);
+        $this->assertConsumesFirst('FooBarBar', $pattern);
+    }
+
+    /**
+     * @test
+     */
+    public function shouldAllowBacktrackingBuilderTemplatePattern()
+    {
+        // given
+        $pattern = Pattern::builder('@Bar')->pattern('FooBar|Foo')->build();
         // when, then
         $this->assertConsumesFirst('FooBar', $pattern);
         $this->assertConsumesFirst('FooBarBar', $pattern);