+++ test/Feature/CleanRegex/_prepared/template/pcre/PatternTest.php
@@ -26,9 +26,9 @@ class PatternTest extends TestCase
     public function templatesWithPlaceholder(): array
     {
         return [
-            'standard'                               => ['#You/her @ her?#', '#You/her X her?#'],
-            'comment (but no "x" flag)'              => ["%You/her #@\n her?%", "%You/her #X\n her?%"],
-            'comment ("x" flag, but also "-x" flag)' => ["%You/her (?x:(?-x:#@\n)) her?%", "%You/her (?x:(?-x:#X\n)) her?%"],
+            'standard'                               => ['#You/her @ her?#', '#You/her (?>X) her?#'],
+            'comment (but no "x" flag)'              => ["%You/her #@\n her?%", "%You/her #(?>X)\n her?%"],
+            'comment ("x" flag, but also "-x" flag)' => ["%You/her (?x:(?-x:#@\n)) her?%", "%You/her (?x:(?-x:#(?>X)\n)) her?%"],
         ];
     }
 
@@ -40,7 +40,7 @@ class PatternTest extends TestCase
         // given
         $pattern = PcrePattern::template('%foo:@%m')->literal('bar%cat');
         // when, then
-        $this->assertPatternIs('%foo:bar\%cat%m', $pattern);
+        $this->assertPatternIs('%foo:(?>bar\%cat)%m', $pattern);
     }
 
     /**
@@ -87,6 +87,6 @@ class PatternTest extends TestCase
         // when
         $pattern = PcrePattern::template("%You/her (?-x:#@\n) her?%x")->literal('X');
         // then
-        $this->assertPatternIs("%You/her (?-x:#X\n) her?%x", $pattern);
+        $this->assertPatternIs("%You/her (?-x:#(?>X)\n) her?%x", $pattern);
     }
 }